Optimise RRDP snapshot and rsync loading.
Check is the object is present by using blob hash _before_ parsing ASN1 and serialising it.
It speeds up second loading of RIPE snapshot.xml from 42s of CPU time to about 1.5s. It will
not happen very often in practice, but it's a reasonable optimisation for snapshot-fallback
and for loading rsync repositories.